+++ src/config/slateToReact/slateDemo.tsx
@@ -1,0 +1,5 @@
+import { Config } from './types'
+
+export const config: Config = {
+  elementTransforms: {},
+}